Added cutadatp module
