+++ multiqc/modules/bowtie2/bowtie_module.py
@@ -18,7 +18,7 @@ class MultiqcModule(BaseMultiqcModule):
 
         # Initialise the parent object
         super(MultiqcModule, self).__init__(name="Bowtie 2", anchor="bowtie2", 
-        target='Bowtie 2',href='http://bowtie-bio.sourceforge.net/bowtie2/', 
+        href='http://bowtie-bio.sourceforge.net/bowtie2/', 
         info="is ultrafast and memory-efficient tool for aligning sequencing"\
                 " reads to long reference sequences.")
